+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Fix
+- fail if `tract` binary not found but `check_same_io_as_tract=True`
 - `Llama` export
 
 ## [0.8.0] - 2023-05-01